feat: Implemented Bubble Sort algorithm on Cannon.
Cannon now fires ever X moves by the player. When it does, it sorts all
of the remaining tiles in ascending order of health (index 0 has the
least health), and fires at the top X tiles in order.

Also added a cannon power bar to indicate when it will fire. The bar
fills up as the player moves, when it reaches the top, the players's
movement is locked and the cannon fires.