fix: Locked camera target Z axis.
Previously, the camera followed the player along the Z axis. This was a
little much. And, now, with the cannon, it can sometimes cause the
camera to tilt so far that the cannon is cut off at the top of the
screen.

This change makes it so that the camera only follows the target along
the X axis (side to side), which is both less disorienting, and keeps
the cannon fully on the screen.