feat: Added line of sight obstruction to cannon.
The cannon can no longer hit targets that are obstructed by walls.

Also, added walls.

Walls are just full-size cubes that are placed randomly around the grid.
The tiles that are to be fired on still show a target marker, but the
turret won't fire on them unless it has an unobstructed line of sight
to it.