feat: Added ability to choose sorting algorithm.
Added a dropdown box to main menu that allows player to choose which
sorting algorithm (Bubble or Insertion) they want the game to use
when sorting targets.

It works by maintaining a static variable that represents the sorting
method. That variable is updated when the dropdown option box is
changed, and checked when the sorting algorithm is called.